Rename graphRequest to graphQuery
