+++ src/network/graphQuery.js
@@ -5,7 +5,7 @@ import client from './graphClient';
 import GraphParser from './graphParser';
 import { TYPE, isValidEnum, getTypeDef } from './graphDataStruct';
 
-class GraphRequest {
+class GraphQuery {
   constructor(queryName, type) {
     this.queryName = queryName;
     this.type = type;
@@ -104,7 +104,7 @@ class GraphRequest {
 * @param orderBy {Object} Object with order by fields. ie. { field: 'blockNum', direction: 'ASC' }
 */
 export function queryAllTopics(filters, orderBy) {
-  const request = new GraphRequest('allTopics', TYPE.topic);
+  const request = new GraphQuery('allTopics', TYPE.topic);
   if (!_.isEmpty(filters)) {
     request.setFilters(filters);
   }
@@ -120,7 +120,7 @@ export function queryAllTopics(filters, orderBy) {
 * @param orderBy {Object} Object with order by fields. ie. { field: 'blockNum', direction: 'DESC' }
 */
 export function queryAllOracles(filters, orderBy) {
-  const request = new GraphRequest('allOracles', TYPE.oracle);
+  const request = new GraphQuery('allOracles', TYPE.oracle);
   if (!_.isEmpty(filters)) {
     request.setFilters(filters);
   }
@@ -134,5 +134,5 @@ export function queryAllOracles(filters, orderBy) {
 * Queries syncInfo from GraphQL.
 */
 export function querySyncInfo() {
-  return new GraphRequest('syncInfo', TYPE.syncInfo).execute();
+  return new GraphQuery('syncInfo', TYPE.syncInfo).execute();
 }
